python_scraper/requirements.txt
fastapi
uvicorn
playwright
motor
python-dotenv